+++ Performance Improvement/src/valhalla/java/com/example/benchmarks/ValhallaValueBenchmark.java
@@ -12,14 +12,17 @@ import java.util.concurrent.TimeUnit;
  *
  * This is {@code ValhallaBenchmark} with ONE change: `record` → `value record`.
  * That single keyword is what actually opts the type into Valhalla's
- * scalarization/flattening — nothing else about the code changes. Method
- * names (sumPointsRecord, computeDistances) intentionally match
- * ValhallaBenchmark so the existing compare/report tooling — which keys rows
- * by method name — lines up "Java 25, identity record" against
- * "Java 28 EA, value record" automatically.
+ * scalarization/flattening — nothing else about the code changes (see
+ * ValhallaBenchmark's Javadoc for why the fields are `short`, not `int`: on
+ * this JVM, nullable-array flattening only engages when the record's field
+ * payload plus its 1-byte null marker fits in 8 bytes, and that constraint
+ * applies equally here). Method names (sumPointsRecord, computeDistances)
+ * intentionally match ValhallaBenchmark so the existing compare/report
+ * tooling — which keys rows by method name — lines up "Java 25, identity
+ * record" against "Java 28 EA, value record" automatically.
  *
  * Expected result: Point[] becomes a genuinely flat, header-free array
- * (~8 bytes/point instead of ~24), so sumPointsRecord/computeDistances go
+ * (~8 bytes/point instead of ~16-24), so sumPointsRecord/computeDistances go
  * from "pointer-chase 5M heap objects" to "scan contiguous memory" — a real,
  * mechanically-explained win, not a JIT hand-wave.
  */
@@ -34,7 +37,7 @@ public class ValhallaValueBenchmark {
     @Param({"1000000", "5000000"})
     private int size;
 
-    value record Point(int x, int y) {}
+    value record Point(short x, short y) {}
 
     /**
      * Array of value-record Points. On Java 28 EA with --enable-preview, the
@@ -45,7 +48,7 @@ public class ValhallaValueBenchmark {
     public long sumPointsRecord() {
         Point[] points = new Point[size];
         for (int i = 0; i < size; i++) {
-            points[i] = new Point(i, i * 2);
+            points[i] = new Point((short) i, (short) (i * 2));
         }
         long sum = 0;
         for (Point p : points) {
@@ -62,7 +65,7 @@ public class ValhallaValueBenchmark {
     public double computeDistances() {
         Point[] points = new Point[size];
         for (int i = 0; i < size; i++) {
-            points[i] = new Point(i % 1000, (i * 7) % 1000);
+            points[i] = new Point((short) (i % 1000), (short) ((i * 7) % 1000));
         }
         double totalDistance = 0.0;
         for (int i = 1; i < size; i++) {
